# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "KYS PORTAL",
    "description": "KYS Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "103.224.246.103:3004",
  "basePath": "/",
...
  "paths": {
...
    "/subscriber": {
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "Create new admin in system",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Admin that we want to create",
            "schema": {
              "$ref": "#/definitions/Subcriber"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/subscriber/getUserProfile/{id}": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "Create new admin in system",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "get info user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/subscriber/uploadProfilePic": {
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "upload Profile Pic",
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "Profile pic of user",
            "type": "file"
          },
          {
            "name": "userID",
            "in": "formData",
            "description": "",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/subscriber/update": {
      "post": {
        "tags": [
          "Subscribers"
        ],
        "description": "Create new admin in system",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Admin that we want to create",
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/subscriber/deleteById": {
      "get": {
        "tags": [
          "Subscribers"
        ],
        "description": "Remove User in system",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New document is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "required": [
        "name",
        "email",
        "password",
        "isSuperAdmin",
        "permissions"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "permissions": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isSuperAdmin": {
          "type": "boolean"
        },
        "personalProfile": {
          "type": "object",
          "salutation": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "fatherName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "placeofBirth": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "religion": {
            "type": "string"
          },
          "caste": {
            "type": "string"
          },
          "motherTongue": {
            "type": "string"
          },
          "dateofBirth": {
            "type": "string"
          },
          "dateofBirthActual": {
            "type": "string"
          },
          "age": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "bloodGroup": {
            "type": "string"
          },
          "maritalStatus": {
            "type": "boolean"
          },
          "weddingAnniversary": {
            "type": "string"
          },
          "differentlyAbled": {
            "type": "string"
          },
          "smoker": {
            "type": "boolean"
          },
          "aadhaarCardNumber": {
            "type": "string"
          },
          "aadhaarCardName": {
            "type": "string"
          },
          "hobbies": {
            "type": "string"
          },
          "empTag": {
            "type": "string"
          }
        },
        "contactDeatils": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
    "UpdateUser": {
      "required": [],
      "properties": {
        "name": {
          "type": "string"
        },
        "_id": {
          "type": "string"
        },
        "permissions": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "isSuperAdmin": {
          "type": "boolean"
        },
        "personalProfile": {
          "type": "object",
          "salutation": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "fatherName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "placeofBirth": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "religion": {
            "type": "string"
          },
          "caste": {
            "type": "string"
          },
          "motherTongue": {
            "type": "string"
          },
          "dateofBirth": {
            "type": "string"
          },
          "dateofBirthActual": {
            "type": "string"
          },
          "age": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "bloodGroup": {
            "type": "string"
          },
          "maritalStatus": {
            "type": "boolean"
          },
          "weddingAnniversary": {
            "type": "string"
          },
          "differentlyAbled": {
            "type": "string"
          },
          "smoker": {
            "type": "boolean"
          },
          "aadhaarCardNumber": {
            "type": "string"
          },
          "aadhaarCardName": {
            "type": "string"
          },
          "hobbies": {
            "type": "string"
          },
          "empTag": {
            "type": "string"
          }
        },
        "contactDeatils": {
          "type": "array",
          "items": {
            "type": "object"
          }
        }
      }
    },
...
    "Subcriber": {
      "required": [
        "name",
        "mobile"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "subcriptions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "email": {
          "type": "string"
        },
        "fatherName": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        },
        "city": {
          "type": "string"
        },
        "mobile": {
          "type": "number"
        },
        "pinCode": {
          "type": "number"
        },
        "state": {
          "type": "number"
        }
      }
    },
...
  }
}